// src/store/user/types.tsx
export enum userActionType {
  SET_TOKEN = "SET_TOKEN",
  CLEAN_TOKEN = "CLEAN_TOKEN",
  SET_USER_INFO = "SET_USER_INFO",
  CLEAN_USER_INFO = "CLEAN_USER_INFO",
  LOGIN_STARTED = "LOGIN_STARTED",
  LOGIN_SUCCESS = "LOGIN_SUCCESS",
  LOGIN_ERROR = "LOGIN_ERROR",
  LOGOUT_SUCCESS = "LOGOUT_SUCCESS",
  SIGNUP_STARTED = "SIGNUP_STARTED",
  SIGNUP_SUCCESS = "SIGNUP_SUCCESS",
  SIGNUP_ERROR = "SIGNUP_ERROR",
  RECOVER_STARTED = "RECOVER_STARTED",
  RECOVER_SUCCESS = "RECOVER_SUCCESS",
  RECOVER_ERROR = "RECOVER_ERROR",
}